Decided to rename references from chatty to beaker to keep things consistent
