+++ packages/client/hmi-client/src/components/llm/tera-beaker-response-thought.vue
