+++ packages/client/hmi-client/vite.config.local.ts
@@ -29,16 +29,16 @@ config.server.proxy = {
 		}`,
 		changeOrigin: true
 	},
-	'/chatty_ws': {
+	'/beaker_ws': {
 		target: `ws://${process.env.jupyter_host || 'jupyter'}:${process.env.jupyter_port || 8888}`,
 		changeOrigin: true,
-		rewrite: (path_str) => path_str.replace(/^\/chatty_ws/, ''),
+		rewrite: (path_str) => path_str.replace(/^\/beaker_ws/, ''),
 		ws: true
 	},
-	'/chatty': {
+	'/beaker': {
 		target: `http://${process.env.jupyter_host || 'jupyter'}:${process.env.jupyter_port || 8888}`,
 		changeOrigin: true,
-		rewrite: (path_str) => path_str.replace(/^\/chatty/, ''),
+		rewrite: (path_str) => path_str.replace(/^\/beaker/, ''),
 		ws: true
 	}
 };
