+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/TGPTResource.java
@@ -18,13 +18,13 @@ public class TGPTResource {
 	@Inject
 	ObjectMapper mapper;
 
-	@ConfigProperty(name = "tgpt.baseUrl", defaultValue = "/chatty/")
+	@ConfigProperty(name = "tgpt.baseUrl", defaultValue = "/beaker/")
 	String baseUrl;
 
-	@ConfigProperty(name = "tgpt.appUrl", defaultValue = "http://localhost:8078/chatty/")
+	@ConfigProperty(name = "tgpt.appUrl", defaultValue = "http://localhost:8078/beaker/")
 	String appUrl;
 
-	@ConfigProperty(name = "tgpt.wsUrl", defaultValue = "ws://localhost:8078/chatty_ws/")
+	@ConfigProperty(name = "tgpt.wsUrl", defaultValue = "ws://localhost:8078/beaker_ws/")
 	String wsUrl;
 
 	@ConfigProperty(name = "tgpt.token", defaultValue = "89f73481102c46c0bc13b2998f9a4fce")
